src/requirements.txt
asn1crypto==1.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
google-api-core==2.1.0
google-auth==2.3.0
google-cloud==0.34.0
google-cloud-appengine-logging==0.1.5
google-cloud-audit-log==0.1.1
google-cloud-core==2.1.0
google-cloud-logging==2.6.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pg8000==1.21.2
proto-plus==1.19.4
protobuf==3.18.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.7.2
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.22
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.6.0
